Refactor chat interface
